requirements.txt
aiofiles==23.2.1
cryptography
fastapi==0.121.0
sqlmodel==0.0.27
uvicorn==0.38.0
python-multipart==0.0.20
PyJWT
